+++ djangocassandra/db/backends/cassandra/utils.py
@@ -1,7 +1,6 @@
 import sys
 
 from functools import (
-    cmp_to_key,
     wraps
 )
